requirements.txt
argcomplete==2.0.0
asgiref==3.5.2
boto3==1.26.20
botocore==1.29.20
certifi==2022.9.24
cfn-flip==1.3.0
charset-normalizer==2.1.1
click==8.1.3
Django==3.2
django-sms==0.5.0
djangorestframework==3.14.0
durationpy==0.5
hjson==3.1.0
idna==3.4
jmespath==1.0.1
kappa==0.6.0
MarkupSafe==2.1.1
placebo==0.9.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-slugify==7.0.0
pytz==2022.6
PyYAML==6.0
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
toml==0.10.2
tqdm==4.64.1
troposphere==4.2.0
urllib3==1.26.13
Werkzeug==2.2.2
zappa==0.56.0
